Fix the ImplicitObjectLowerBound canonicalizer rule
