+++ src/canonicalizer/canonicalizer.cc
@@ -20,6 +20,7 @@
 #include "rules/equal_numeric_bounds_as_const.h"
 #include "rules/if_without_then_else.h"
 #include "rules/implicit_array_lower_bound.h"
+#include "rules/implicit_object_lower_bound.h"
 #include "rules/max_contains_without_contains.h"
 #include "rules/min_contains_without_contains.h"
 #include "rules/null_as_const.h"
@@ -44,6 +45,7 @@ sourcemeta::jsonbinpack::Canonicalizer::Canonicalizer(
   this->bundle.template add<EqualNumericBoundsAsConst>();
   this->bundle.template add<IfWithoutThenElse>();
   this->bundle.template add<ImplicitArrayLowerBound>();
+  this->bundle.template add<ImplicitObjectLowerBound>();
   this->bundle.template add<MaxContainsWithoutContains>();
   this->bundle.template add<MinContainsWithoutContains>();
   this->bundle.template add<NullAsConst>();
